requirements.txt
asgiref==3.8.1
boto3==1.34.84
botocore==1.34.84
colorama==0.4.6
coverage==7.4.4
Django==5.0.4
django-storages==1.14.2
django-treebeard==4.7.1
djangorestframework==3.15.1
exceptiongroup==1.2.1
factory-boy==3.3.0
Faker==24.11.0
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
PyJWT==2.8.0
pymemcache==4.0.0
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
zstd==1.5.5.1
